requirements.txt
flask
flask-sqlalchemy
flask-wtf
sqlite3
argon2-cffi